Fix magic mirror prediction
